refactor: alter the controller to include the responseEntity to costum request from users

+++ src/main/java/dev/Braz/spring_1/Player/PlayerController.java
@@ -1,8 +1,11 @@
 package dev.Braz.spring_1.Player;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/player")
@@ -20,27 +23,41 @@ public class PlayerController {
     }
 
     @GetMapping("/list")
-    public List<PlayerDTO> ListPlayers(){
-        return playerService.listPlayers();
+    public ResponseEntity<List<PlayerDTO>> ListPlayers(){
+        List<PlayerDTO> playerList =  playerService.listPlayers();
+        return ResponseEntity.ok(playerList);
     }
 
     @PostMapping("/create")
-    public PlayerDTO CreatePlayer(@RequestBody PlayerDTO player){
-        return playerService.CreatePlayer(player);
+    public ResponseEntity<?> CreatePlayer(@RequestBody PlayerDTO player){
+        PlayerDTO NewPlayer = playerService.CreatePlayer(player);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("The Player was created! \n" + NewPlayer);
     }
 
     @GetMapping("/list/{id}")
-    public PlayerDTO ListById(@PathVariable Long id) {
-        return playerService.listPlayerById(id);
+    public ResponseEntity<?> ListById(@PathVariable Long id) {
+       PlayerDTO player = playerService.listPlayerById(id);
+        if(player != null)
+           return ResponseEntity.ok(player);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("The id: " + id+ " was not found!" );
     }
 
     @PutMapping("/update/{id}")
-    public PlayerDTO UpdatePlayerById(@PathVariable Long id, @RequestBody PlayerDTO updated){
-        return playerService.UpdatePlayerById(id, updated);
+    public ResponseEntity<?> UpdatePlayerById(@PathVariable Long id, @RequestBody PlayerDTO updated){
+        PlayerDTO updatedPlayer = playerService.UpdatePlayerById(id, updated);
+        if(updatedPlayer != null)
+            return ResponseEntity.ok(updatedPlayer);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("The player with id: "+ id +  "was not found");
     }
 
     @DeleteMapping("/delete/{id}")
-    public void DeletedPlayerById(@PathVariable Long id){
-        playerService.DeletePlayerById(id);
+    public ResponseEntity<String> DeletedPlayerById(@PathVariable Long id){
+        if(playerService.listPlayerById(id) != null)
+        {
+            playerService.DeletePlayerById(id);
+            return ResponseEntity.ok(" id: " + id +" Player deleted");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("The player with id: "+ id + " was not found");
     }
 }
